remote_billing_page: Extract method to get RemoteBillingIdentityDict.
This will likely be used in other files.
